oh_mary_cmm_analysis/requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Data collection
praw>=7.7.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP & Analysis (simplified for Mac compatibility)
scikit-learn>=1.3.0
nltk>=3.8

# Statistical modeling & causal inference
statsmodels>=0.14.0
scipy>=1.11.0
linearmodels>=5.3  # Panel regression with entity/time fixed effects

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data formats
pyarrow>=14.0.0  # For reading/writing Parquet files

# Public signals collection
pytrends>=4.9.0  # Google Trends API
wikipediaapi>=0.6.0  # Wikipedia API
playwright>=1.40.0  # Headless browser for TikTok/Instagram
tenacity>=8.2.0  # Retry logic with backoff

# Utilities
tqdm>=4.65.0
